test(hub): give hub-manager its own spec file
